classification-worker/requirements.txt
flask
flask-cors
mongo
gunicorn
email
